% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head-block.R
\name{new_head_block}
\alias{new_head_block}
\title{Head block}
\usage{
new_head_block(data, n_rows = numeric())
}
\arguments{
\item{data}{Tabular data to filter (rows)}

\item{n_rows}{Number of rows to return.}
}
\description{
An example of a "head" block intendet to show the first n rows a a table.
}
